Fix failing proptest for HighLatencyOp by assuming no overflow
The `IMonotonicClock::sleep` signature may need some adjustments in
light of the realization that adding a `std::time::Instant` and
`std::time::Duration` could overflow in a platform-specific way.

The docs for `std::time::Duration::MAX` indicate the range of
`std::time::Instant`s are at least an order of magnitude longer than
the time since the big bang. Assuming none of the supported platform's
clocks are relatively near the end of their `Instant` range for all
conceivable practical purposes overflow is a non-issue.